SCRUM-4185 added service, validator, and dao for sqtrGeneAssociations
